Clean marketStatsWithOrder
